backup: Ensure prompt is bounded to user actions.
In this commit we decouple between the backup error, which
shows the warning icon and the backup prompts which will now
be shown only after a user has trigerred an operation that needs
backup: "sent payment", "received payment", "add funds".

To implement it we add another stream controller in the backup
bloc that recevies both the backup error and the user operations
and emit an event only when needed.