ctp: fix layout issue.
This comit fixes a layout issue that was discovered in connect to pay
screen (https://github.com/breez/breezmobile/issues/88).
From some reason, bellow a certiain height resolution the Pay button
wasn't rendered.
The fix is to show this button anyway and make sure we use scroll view
for the case the user device height is is not enough to show all
elements.