refactor send payments.
The main purpose of this PR is to be able to retry a payment attempt.
More specifically we have noticed that one of the major failure reasons
regarding payments is not synched network graph.
Here we attempt to detect such failures, wait for the sync job to
complete and then retry the payment.