Fix search path for ios build.
Search paths are taken in the order they are written so we can't
use them as a mechanism to choose the target. The user should
copy the framework related to his specific architecture to the
binding.framework folder. This enable the automatic build also
to use this mechanism when building for release.